Moved people controller out of api folder
I wanted to put the controller in the api namespace
but since all the devise stuff is actually hooked up
I decided to keep it in the space that wasn’t already protected
by login.

First test passing. Next to create a service layer to fetch the results 
from the developer api.